+++ qgreenland/constants.py
@@ -1,12 +1,10 @@
 import os
 from enum import Enum
 
-import yamale
-
 from qgreenland import __version__
+from qgreenland.util.config import make_config
 
 PROJECT = 'qgreenland'
-
 
 ENVIRONMENT = os.environ.get('ENVIRONMENT', 'dev')
 
@@ -15,13 +13,14 @@ DATA_DIR = '/luigi/data'
 RELEASES_DIR = f'{DATA_DIR}/release'
 WIP_DIR = f'{DATA_DIR}/luigi-wip'
 ASSETS_DIR = f'{PACKAGE_DIR}/assets'
-CONFIG_DIR = f'{PACKAGE_DIR}/config'
-CONFIG_SCHEMA_DIR = f'{CONFIG_DIR}/schema'
-
 if 'dev' in __version__:
     RELEASE_DIR = f'{RELEASES_DIR}/dev/{__version__}'
 else:
     RELEASE_DIR = f'{RELEASES_DIR}/{__version__}'
+
+CONFIG_DIR = f'{PACKAGE_DIR}/config'
+CONFIG_SCHEMA_DIR = f'{CONFIG_DIR}/schema'
+CONFIG = make_config(config_dir=CONFIG_DIR, schema_dir=CONFIG_SCHEMA_DIR)
 
 # TMP_DIR is the same as WIP_DIR because os.rename doesn't allow cross-mount
 # renaming. Make it a subdir?
@@ -47,37 +46,6 @@ PROJECT_CRS = 'EPSG:3411'
 
 # URS stuff
 URS_COOKIE = 'urs_user_already_logged'
-
-
-def _load_config(config_filename):
-    """Validate config file against schema with Yamale.
-
-    It is expected that the given config filename in CONFIG_DIR has a schema of
-    matching name in CONFIG_SCHEMA_DIR.
-
-    Yamale can read in directories of config files, so it returns a list of
-    (data, fp) tuples. We always read single files, so we return just the data
-    from result[0][0].
-    """
-    config_fp = os.path.join(CONFIG_DIR, config_filename)
-    schema_fp = os.path.join(CONFIG_SCHEMA_DIR, config_filename)
-
-    if not os.path.isfile(config_fp):
-        return NotImplementedError(
-            'Loading is supported for only one config file at a time.'
-        )
-
-    schema = yamale.make_schema(schema_fp)
-    config = yamale.make_data(config_fp)
-    yamale.validate(schema, config)
-
-    return config[0][0]
-
-
-CONFIG = {
-    'layers': _load_config('layers.yml'),
-    'layer_groups': _load_config('layer_groups.yml')
-}
 
 
 class TaskType(Enum):