+++ qgreenland/tasks/common/raster.py
@@ -9,10 +9,10 @@ from osgeo import gdal
 from shapely.geometry import Polygon
 
 from qgreenland.constants import BBOX_POLYGON, PROJECT_CRS, TaskType
-from qgreenland.util.luigi import LayerConfigMixin
+from qgreenland.util.luigi import LayerTask
 
 
-class BuildRasterOverviews(LayerConfigMixin, luigi.Task):
+class BuildRasterOverviews(LayerTask):
     task_type = TaskType.WIP
     requires_task = luigi.Parameter()
 
@@ -20,8 +20,7 @@ class BuildRasterOverviews(LayerConfigMixin, luigi.Task):
         return self.requires_task
 
     def output(self):
-        fn = os.path.basename(self.input().path)
-        of = os.path.join(self.outdir, 'overviews', fn)
+        of = os.path.join(self.outdir, 'overviews', self.filename)
         return luigi.LocalTarget(of)
 
     def run(self):
@@ -53,7 +52,7 @@ class BuildRasterOverviews(LayerConfigMixin, luigi.Task):
                 ds.build_overviews(overview_levels, resampling_method)
 
 
-class ReprojectRaster(LayerConfigMixin, luigi.Task):
+class ReprojectRaster(LayerTask):
     task_type = TaskType.WIP
     requires_task = luigi.Parameter()
 
@@ -61,8 +60,7 @@ class ReprojectRaster(LayerConfigMixin, luigi.Task):
         return self.requires_task
 
     def output(self):
-        fn = os.path.basename(self.input().path)
-        of = os.path.join(self.outdir, 'reproject', fn)
+        of = os.path.join(self.outdir, 'reproject', self.filename)
         return luigi.LocalTarget(of)
 
     def run(self):
@@ -77,7 +75,7 @@ class ReprojectRaster(LayerConfigMixin, luigi.Task):
                       **warp_kwargs)
 
 
-class SubsetRaster(LayerConfigMixin, luigi.Task):
+class SubsetRaster(LayerTask):
     task_type = TaskType.WIP
     requires_task = luigi.Parameter()
 
@@ -85,8 +83,7 @@ class SubsetRaster(LayerConfigMixin, luigi.Task):
         return self.requires_task
 
     def output(self):
-        fn = os.path.basename(self.input().path)
-        of = os.path.join(self.outdir, 'subset', fn)
+        of = os.path.join(self.outdir, 'subset', self.filename)
         return luigi.LocalTarget(of)
 
     def run(self):