+++ qgreenland/tasks/common/shapefile.py
@@ -4,7 +4,7 @@ import zipfile
 import luigi
 
 from qgreenland.constants import TaskType
-from qgreenland.util.luigi import LayerConfigMixin
+from qgreenland.util.luigi import LayerTask
 from qgreenland.util.misc import temporary_path_dir
 from qgreenland.util.shapefile import (find_shapefile_in_dir,
                                        reproject_shapefile,
@@ -14,7 +14,7 @@ from qgreenland.util.shapefile import (find_shapefile_in_dir,
 # TODO: Is there any task history? e.g. can we look at the final output target
 # and generate a list of tasks that were performed to generate it?
 
-class UnzipShapefile(LayerConfigMixin, luigi.Task):
+class UnzipShapefile(LayerTask):
     task_type = TaskType.WIP
     requires_task = luigi.Parameter()
 
@@ -33,7 +33,7 @@ class UnzipShapefile(LayerConfigMixin, luigi.Task):
             zf.close()
 
 
-class ReprojectShapefile(LayerConfigMixin, luigi.Task):
+class ReprojectShapefile(LayerTask):
     task_type = TaskType.WIP
     requires_task = luigi.Parameter()
 
@@ -41,18 +41,24 @@ class ReprojectShapefile(LayerConfigMixin, luigi.Task):
         return self.requires_task
 
     def output(self):
-        return luigi.LocalTarget(f'{self.outdir}/reproject/')
+        # TODO: DRY this out. Place responsibility in LayerTask.
+        out_dir = os.path.join(
+            self.outdir,
+            'reproject',
+            self.id
+        )
+        return luigi.LocalTarget(out_dir)
 
     def run(self):
         shapefile = find_shapefile_in_dir(self.input().path)
 
         gdf = reproject_shapefile(shapefile)
         with temporary_path_dir(self.output()) as temp_path:
-            fn = os.path.join(temp_path, 'shapefile.shp')
+            fn = os.path.join(temp_path, self.filename)
             gdf.to_file(fn, driver='ESRI Shapefile')
 
 
-class SubsetShapefile(LayerConfigMixin, luigi.Task):
+class SubsetShapefile(LayerTask):
     task_type = TaskType.WIP
     requires_task = luigi.Parameter()
 
@@ -67,5 +73,5 @@ class SubsetShapefile(LayerConfigMixin, luigi.Task):
 
         gdf = subset_shapefile(shapefile)
         with temporary_path_dir(self.output()) as temp_path:
-            fn = os.path.join(temp_path, 'shapefile.shp')
+            fn = os.path.join(temp_path, self.filename)
             gdf.to_file(fn, driver='ESRI Shapefile')