+++ scripts/cleanup.sh
@@ -1,2 +1,2 @@
 docker-compose exec luigi python -c \
-  'from qgreenland.util.misc import cleanup_output_dirs; cleanup_output_dirs()'
+  'from qgreenland.util.misc import cleanup_output_dirs; cleanup_output_dirs(delete_fetch_dir=False)'