+++ scripts/query_cmr_granules.sh
@@ -1,0 +1,1 @@
+python -c "from qgreenland.util.cmr import pretty_search_cmr_granules; pretty_search_cmr_granules(short_name='$1', version='$2')"